PrintPreview changes:
When you set the print plugin flag 'appendLegendOptions' to true, you can manage these legend options in the PrintPreview:

* Icon size
* Font family
* Font style
* Font size
* Force label
* Dpi

Playng with these options you can fix common issues as too big legend in the printed map or incorrect font family.

Also, you have an option called 'addGraticuleControl' for add a new checkbox into the the print preview to activate or deactivate the graticule in the print. When you enable this option you can add another config parameter to the print plugin called 'offsetByScale' that fixes the graticule layer labels to show it into the printed PDF.