Geocoder changes.
Return the selected area by return type:

* Current extent by default
* Geometry Union (for geocoder selection)
* Administrative Area List as 'Geoname_PARENT,Geoname2_PARENT2' (for geocoder selection)
* Administrative Area Subs as 'Geoname_PARENT' (for geocoder selection)

You must use the method `getSelectedAreas` to obtain the selected area in the selected format.

When you select the geocoder selector in the combobox and the parameter `selectReturnType` is setted to true, it shows another combobox to select the return type.

Add italian and eng translations.